Changes to reader to make it easier to customise behaviour.
The main approach used to is to make an iterable class (using the TypedResourceIterable interface), the big improvement is handing of validation errors, previously a validation error during processing would cause the iterator to raise a StopIteration even if the ValidationError was caught (the expected behaviour of yield) however this would mean that if you wanted to process an entire file and report bad rows.

The only semi breaking change is dropping the *args param for passing additional params to the underlying csv.reader, I would strongly recommend using the **kwargs param anyway as it is more explicit.